erros

+++ jogos/views.py
@@ -1,6 +1,7 @@
 from sanic.views import HTTPMethodView
 from sanic.response import json
 from jogos.models import Jogos
+from sanic.exceptions import NotFound
 
 class ListarJogos(HTTPMethodView):
     async def get(self, request):
@@ -15,18 +16,28 @@ class ListarJogos(HTTPMethodView):
 
 class ListarJogo(HTTPMethodView):
     async def get(self, request, pk):
-        return json(dict(await Jogos.get(pk=pk)))
+        jogo = await Jogos.get_or_none(pk=pk)
+        if jogo is None:
+            raise NotFound('Id not found')
+        else:
+            return json(dict(jogo))
 
     async def put(self, request, pk):
-        jogo = await Jogos.get(pk=pk)
-        await jogo.update_from_dict({
-            'nome': request.json['nome'],
-            'categoria': request.json['categoria'],
-            'console': request.json['console']
-        }).save()
-        return json({'message': f'Jogo {jogo.nome} atualizado'})
+        jogo = await Jogos.get_or_none(pk=pk)
+        if jogo is not None:
+            await jogo.update_from_dict({
+                'nome': request.json['nome'],
+                'categoria': request.json['categoria'],
+                'console': request.json['console']
+            }).save()
+            return json({'message': f'Jogo {jogo.nome} atualizado'})
+        else:
+            raise NotFound('Id not found')
 
     async def delete(self, request, pk):
         jogo = await Jogos.get(pk=pk)
-        await jogo.delete()
-        return json({'message': f'Jogo {jogo.nome} removido'})
+        if jogo is not None:
+            await jogo.delete()
+            return json({'message': f'Jogo {jogo.nome} removido'})
+        else:
+            raise NotFound('Id not found')
